Added raycasting based on Bresenham's algorithm. WIP, will need reformatting so scalable pixel data gets fetched from the images which then get drawn onto a view plane.

+++ src/game/level_change.c
@@ -17,8 +17,8 @@
  */
 void	init_level_params(t_level *level, t_player *player)
 {
-	player->x = (double)level->player_x + 0.5;
-	player->y = (double)level->player_y + 0.5;
+	player->x = level->player_x + 0.5;
+	player->y = level->player_y + 0.5;
 	if (level->direction == 'N')
 		player->angle = PI / 2.0;
 	else if (level->direction == 'E')
